feat: retry on errors
Retry on ConnectTimeout, ReadTimeout, or random Bad Gateways (502s) from the Steam API.

The program will retry up to 5 times, with a 5 second delay. If there are still errors, execution is paused with a prompt asking the user if they should continue.

This should reduce the amount of times I have to restart the program when Steam decides to shit itself 🙃